[qca-nss-dp]: Support ptp service code for timestamp function
In RX path, the packet with the ptp service code is received,
this packet will be delivered to the phy driver api rxtstamp
for handing rx timestamping.

In TX path, the socket option SKBTX_HW_TSTAMP is enabled, the
sending packet will be delivered to the phy driver api txtstamp
for handing tx timestamping.

Change-Id: I373997d347ad52d7793367eed88866e1c93ef26d
Signed-off-by: Luo Jie <luoj@codeaurora.org>