[qca-nss-dp] Enable interrupts only after NAPI enable.
Currently interrupts are getting enabled in the probe function.
In the interrupt handler, NAPI is getting scheduled after
disabling the EDMA interrupts. NAPI is getting enabled in the
netdev's open function. NAPI handler is processing the packets
and enabling the interrupts back.

So, if some packet comes before the netdev's open function, the
interrupt handler will disable the interrupts and schedule the
NAPI handler (but handler will not get invoked since NAPI is
not yet enabled). So, EDMA hardware will never send any further
interrupts to the host driver.

This change is enabling the interrupts after NAPI is getting
enabled.

Signed-off-by: Manish Verma <maniverm@codeaurora.org>
Change-Id: I2cd43172da445567d5ad66d3a511176afb62c54b