nss_dp_main: Use a 'phy-handle' property to connect to the PHY
The original method of connecting a PHY to the ethernet controller
requires the "qcom,link-poll", and "qcom,phy-mdio-addr" devicetree
properties. This is redundant. The PHY node already contains the MDIO
address, and attaching a PHY implies "link-poll".

Allow using a "phy-handle" property, in which case ignore the
aforementioned "qcom,*" properties. These can now be considered
deprecated, but are not removed yet, for backwards-compatibility.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>