ethtools: Constify nss_dp_ethtool_ops struct
The nss_dp_ethtool_ops is placed in the ".data" section. Thus a simple
pointer overshoot might overwrite the ops structure. Given the quality
of this driver, that would never happen.

As a precaution, declare the structure "const". With this simple
modification, it now end up in the ".rodata" section.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>